+++ inc/customizer/controls/class-Customizer_Multiselect.php
@@ -8,7 +8,7 @@
  * @copyright  2015 WebMan - Oliver Juhas
  *
  * @since    1.0
- * @version  1.0
+ * @version  1.6.0
  */
 
 
@@ -26,13 +26,13 @@ class Receptar_Customizer_Multiselect extends WP_Customize_Control {
 				<span class="customize-control-title"><?php echo $this->label; ?></span>
 				<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo $this->description; ?></span><?php endif; ?>
 
-				<select name="<?php echo $this->id; ?>" multiple="multiple" <?php $this->link(); ?>>
+				<select name="<?php echo esc_attr( $this->id ); ?>" multiple="multiple" <?php $this->link(); ?>>
 
 					<?php
 
 					foreach ( $this->choices as $value => $name ) {
 
-						echo '<option value="' . $value . '" ' . selected( $this->value(), $value, false ) . '>' . $name . '</option>';
+						echo '<option value="' . esc_attr( $value ) . '" ' . selected( $this->value(), $value, false ) . '>' . $name . '</option>';
 
 					}
 
@@ -47,5 +47,3 @@ class Receptar_Customizer_Multiselect extends WP_Customize_Control {
 	}
 
 } // /Receptar_Customizer_Multiselect
-
-?>