+++ inc/customizer/controls/class-Customizer_Select.php
@@ -8,7 +8,7 @@
  * @copyright  2015 WebMan - Oliver Juhas
  *
  * @since    1.0
- * @version  1.0
+ * @version  1.6.0
  */
 
 
@@ -25,7 +25,7 @@ class Receptar_Customizer_Select extends WP_Customize_Control {
 			<label>
 				<span class="customize-control-title"><?php echo $this->label; ?></span>
 				<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo $this->description; ?></span><?php endif; ?>
-				<select name="<?php echo $this->id; ?>" <?php $this->link(); ?>>
+				<select name="<?php echo esc_attr( $this->id ); ?>" <?php $this->link(); ?>>
 
 					<?php
 
@@ -38,7 +38,7 @@ class Receptar_Customizer_Select extends WP_Customize_Control {
 						} elseif ( 0 === strpos( $value, '/optgroup' ) ) {
 							echo '</optgroup>';
 						} else {
-							echo '<option value="' . $value . '" ' . selected( $this->value(), $value, false ) . '>' . $name . '</option>';
+							echo '<option value="' . esc_attr( $value ) . '" ' . selected( $this->value(), $value, false ) . '>' . $name . '</option>';
 						}
 
 					}
@@ -53,5 +53,3 @@ class Receptar_Customizer_Select extends WP_Customize_Control {
 	}
 
 } // /Receptar_Customizer_Select
-
-?>